requirements.txt
langchain
langchain-openai
pyhive
thrift
thrift-sasl
python-dotenv